Refactor constants into package
